Day 1
Created the site, the form and the list, implemented some features. (add a new value on the to-do list, on page load it renders the list by reaching for it on the localStorage variable called todo).
Working on the feature to mark as done, if marked must go to the end of the list, if unmarked must go to the top